+++ src/middlewares/insights.ts
